requirements.txt
cffi==1.14.5
cryptography==3.4.7
numpy==1.21.0
Pillow==8.2.0
pycparser==2.20
